Debug in process
